Changed output to MixingProvider

+++ DreamingPhoenix/DreamingPhoenix/Converter/FloatToSecondsConverter.cs
@@ -10,7 +10,7 @@ namespace DreamingPhoenix.Converter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            return Math.Round(((float)value) / 1000, 1).ToString() + " sec";
+            return Math.Round(((double)value) / 1000, 1).ToString() + " sec";
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
